requirements.txt
Django==1.9.5
django-dia==0.3
django-rest-swagger==0.3.6
djangorestframework==3.3.3
Markdown==2.6.6
PyYAML==3.11
six==1.10.0
